fix: address evm charge review feedback
