--- .changeset/native-evm-authorization.md
@@ -1,5 +1,0 @@
----
-'mppx': patch
----
-
-Added native Payment-auth EVM charge authorization support.
