+++ src/middlewares/hono.test.ts
@@ -230,23 +230,25 @@ describe('charge', () => {
         }),
         evm_server.charge({
           currency: evm_server.assets.baseSepolia.USDC,
-          facilitator: {
-            async verify(paymentPayload: PaymentPayload) {
-              return {
-                isValid: true,
-                payer: payerOf(paymentPayload),
-              }
-            },
-            async settle(paymentPayload: PaymentPayload) {
-              return {
-                network: paymentPayload.accepted.network,
-                payer: payerOf(paymentPayload),
-                success: true,
-                transaction,
-              }
+          recipient: accounts[0].address,
+          x402Options: {
+            facilitator: {
+              async verify(paymentPayload: PaymentPayload) {
+                return {
+                  isValid: true,
+                  payer: payerOf(paymentPayload),
+                }
+              },
+              async settle(paymentPayload: PaymentPayload) {
+                return {
+                  network: paymentPayload.accepted.network,
+                  payer: payerOf(paymentPayload),
+                  success: true,
+                  transaction,
+                }
+              },
             },
           },
-          recipient: accounts[0].address,
         }),
       ],
       secretKey,
